+++ ESP/Settings/VisualWidth.cs
@@ -8,7 +8,7 @@ public partial class Settings
   public static ConfigEntry<int> configStructureCollidersLineWidth;
   public static ConfigEntry<int> configDestructibleCollidersLineWidth;
   public static ConfigEntry<int> configCreatureFireLineWidth;
-  public static ConfigEntry<int> configTrackedCreatureRayWidth;
+  public static ConfigEntry<int> configTrackedObjectRayWidth;
   public static ConfigEntry<int> configBreedingLineWidth;
   public static ConfigEntry<int> configCoverRayWidth;
   public static ConfigEntry<int> configPlayerCoverRayWidth;
@@ -79,8 +79,8 @@ public partial class Settings
     OnWidthChanged(configPlayerCoverRayWidth, Tag.PlayerCoverBlocked);
     configSmokeLineWidth = config.Bind(section, "Smoke", 2, "");
     OnWidthChanged(configSmokeLineWidth, Tag.Smoke);
-    configTrackedCreatureRayWidth = config.Bind(section, "Tracted creature rays", 10, "");
-    OnWidthChanged(configTrackedCreatureRayWidth, Tag.TrackedCreature);
+    configTrackedObjectRayWidth = config.Bind(section, "Tracted object rays", 10, "");
+    OnWidthChanged(configTrackedObjectRayWidth, Tag.TrackedObject);
     configSpawntPointOneTimeRayWidth = config.Bind(section, "Spawn points (one time)", 2, "");
     OnWidthChanged(configSpawntPointOneTimeRayWidth, Tag.SpawnPointOneTime);
     configSpawntPointRespawningRayWidth = config.Bind(section, "Spawn points (respawning)", 2, "");